pyproject.toml
[tool.poetry]
name = "web-auth-sdk"
version = "0.1.0"
description = "authorization SDK used to build protected Web APIs"
authors = ["yangaound@gmail.com"]
readme = "README.md"
repository = "https://github.com/yangaound/web-auth-sdk"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Plugins",
    "Environment :: Web Environment",
    "Framework :: FastAPI",
    "Framework :: Django",
    "Framework :: Flask",
    "Operating System :: POSIX",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
]
packages = [
    { include = "web_auth" },
]
[tool.poetry.dependencies]
python = ">=3.9.0,<=3.10"
PyJWT = ">=2.2.0,<3.0"
requests = ">=2.24.0,<3.0"
petl = "1.0.0"

[tool.poetry.dev-dependencies]
fastapi = "^0.81.0"
Flask = "^2.2.2"
pytest = "^7.1.3"
autoflake = "^1.4"
black = "^22.6.0"
isort = "^5.10.1"
pylint = "^2.14.5"
pylint-quotes = "^0.2.3"
pytest-asyncio = "^0.19.0"
tox = "^3.25.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
skip-string-normalization = true

[tool.isort]
profile = "black"
line_length = 120

[tool.pylint.format]
max-line-length=120
ignore-long-lines = "^\\s*(# )"
disable=[
    "R",
    "arguments-differ",
    "broad-except",
    "fixme",
    "invalid-name",
    "logging-fstring-interpolation",
    "missing-class-docstring",
    "missing-function-docstring",
    "missing-module-docstring",
    "no-self-argument",
    "pointless-statement",
    "protected-access",
    "raise-missing-from",
    "singleton-comparison",
    "super-init-not-called",
    "too-many-lines",
    "unidiomatic-typecheck",
    "unnecessary-lambda",
    "unused-argument",
    "unused-import",
    "import-outside-toplevel",
    "consider-using-f-string",
    "useless-parent-delegation",
    "keyword-arg-before-vararg",
    "redefined-outer-name",
]
